Parte 2 paso 30

+++ app/src/main/java/net/iessochoa/tomassolerlinares/practica5/repository/DiarioRepository.java
@@ -50,7 +50,7 @@ public class DiarioRepository {
     }
 
 
-    public Single<Float> getValoracionTotal() {
+    public Single<Integer> getValoracionTotal() {
         return mdiarioDao.getValoracionTotal();
     }
